+++ Sources/NIO/BSDSocketAPICommon.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the SwiftNIO open source project
 //
-// Copyright (c) 2020 Apple Inc. and the SwiftNIO project authors
+// Copyright (c) 2020-2021 Apple Inc. and the SwiftNIO project authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
@@ -78,13 +78,7 @@ internal enum Shutdown: _SocketShutdownProtocol {
     case RDWR
 }
 
-public enum NIOBSDSocket {
-#if os(Windows)
-    public typealias Handle = SOCKET
-#else
-    public typealias Handle = CInt
-#endif
-
+extension NIOBSDSocket {
 #if os(Windows)
     internal static let invalidHandle: Handle = INVALID_SOCKET
 #else
@@ -109,112 +103,6 @@ extension NIOBSDSocket.SocketType: Equatable {
 extension NIOBSDSocket.SocketType: Hashable {
 }
 
-extension NIOBSDSocket {
-    /// Specifies the addressing scheme that the socket can use.
-    public struct AddressFamily: RawRepresentable {
-        public typealias RawValue = CInt
-        public var rawValue: RawValue
-        public init(rawValue: RawValue) {
-            self.rawValue = rawValue
-        }
-    }
-}
-
-extension NIOBSDSocket.AddressFamily: Equatable {
-}
-
-extension NIOBSDSocket.AddressFamily: Hashable {
-}
-
-extension NIOBSDSocket {
-    /// Specifies the type of protocol that the socket can use.
-    public struct ProtocolFamily: RawRepresentable {
-        public typealias RawValue = CInt
-        public var rawValue: RawValue
-        public init(rawValue: RawValue) {
-            self.rawValue = rawValue
-        }
-    }
-}
-
-extension NIOBSDSocket.ProtocolFamily: Equatable {
-}
-
-extension NIOBSDSocket.ProtocolFamily: Hashable {
-}
-
-extension NIOBSDSocket {
-    /// Defines socket option levels.
-    public struct OptionLevel: RawRepresentable {
-        public typealias RawValue = CInt
-        public var rawValue: RawValue
-        public init(rawValue: RawValue) {
-            self.rawValue = rawValue
-        }
-    }
-}
-
-extension NIOBSDSocket.OptionLevel: Equatable {
-}
-
-extension NIOBSDSocket.OptionLevel: Hashable {
-}
-
-extension NIOBSDSocket {
-    /// Defines configuration option names.
-    public struct Option: RawRepresentable {
-        public typealias RawValue = CInt
-        public var rawValue: RawValue
-        public init(rawValue: RawValue) {
-            self.rawValue = rawValue
-        }
-    }
-}
-
-extension NIOBSDSocket.Option: Equatable {
-}
-
-extension NIOBSDSocket.Option: Hashable {
-}
-
-// Address Family
-extension NIOBSDSocket.AddressFamily {
-    /// Address for IP version 4.
-    public static let inet: NIOBSDSocket.AddressFamily =
-            NIOBSDSocket.AddressFamily(rawValue: AF_INET)
-
-    /// Address for IP version 6.
-    public static let inet6: NIOBSDSocket.AddressFamily =
-            NIOBSDSocket.AddressFamily(rawValue: AF_INET6)
-
-    /// Unix local to host address.
-    public static let unix: NIOBSDSocket.AddressFamily =
-            NIOBSDSocket.AddressFamily(rawValue: AF_UNIX)
-}
-
-// Protocol Family
-extension NIOBSDSocket.ProtocolFamily {
-    /// IP network 4 protocol.
-    public static let inet: NIOBSDSocket.ProtocolFamily =
-            NIOBSDSocket.ProtocolFamily(rawValue: PF_INET)
-
-    /// IP network 6 protocol.
-    public static let inet6: NIOBSDSocket.ProtocolFamily =
-            NIOBSDSocket.ProtocolFamily(rawValue: PF_INET6)
-
-    /// UNIX local to the host.
-    public static let unix: NIOBSDSocket.ProtocolFamily =
-            NIOBSDSocket.ProtocolFamily(rawValue: PF_UNIX)
-}
-
-#if !os(Windows)
-    extension NIOBSDSocket.ProtocolFamily {
-        /// UNIX local to the host, alias for `PF_UNIX` (`.unix`)
-        public static let local: NIOBSDSocket.ProtocolFamily =
-                NIOBSDSocket.ProtocolFamily(rawValue: PF_LOCAL)
-    }
-#endif
-
 // Socket Types
 extension NIOBSDSocket.SocketType {
     /// Supports datagrams, which are connectionless, unreliable messages of a
@@ -238,68 +126,8 @@ extension NIOBSDSocket.SocketType {
     #endif
 }
 
-// Option Level
-extension NIOBSDSocket.OptionLevel {
-    /// Socket options that apply only to IP sockets.
-    #if os(Linux) || os(Android)
-        public static let ip: NIOBSDSocket.OptionLevel =
-                NIOBSDSocket.OptionLevel(rawValue: CInt(IPPROTO_IP))
-    #else
-        public static let ip: NIOBSDSocket.OptionLevel =
-                NIOBSDSocket.OptionLevel(rawValue: IPPROTO_IP)
-    #endif
-
-    /// Socket options that apply only to IPv6 sockets.
-    #if os(Linux) || os(Android)
-        public static let ipv6: NIOBSDSocket.OptionLevel =
-                NIOBSDSocket.OptionLevel(rawValue: CInt(IPPROTO_IPV6))
-    #elseif os(Windows)
-        public static let ipv6: NIOBSDSocket.OptionLevel =
-                NIOBSDSocket.OptionLevel(rawValue: IPPROTO_IPV6.rawValue)
-    #else
-        public static let ipv6: NIOBSDSocket.OptionLevel =
-                NIOBSDSocket.OptionLevel(rawValue: IPPROTO_IPV6)
-    #endif
-
-    /// Socket options that apply only to TCP sockets.
-    #if os(Linux) || os(Android)
-        public static let tcp: NIOBSDSocket.OptionLevel =
-                NIOBSDSocket.OptionLevel(rawValue: CInt(IPPROTO_TCP))
-    #elseif os(Windows)
-        public static let tcp: NIOBSDSocket.OptionLevel =
-                NIOBSDSocket.OptionLevel(rawValue: IPPROTO_TCP.rawValue)
-    #else
-        public static let tcp: NIOBSDSocket.OptionLevel =
-                NIOBSDSocket.OptionLevel(rawValue: IPPROTO_TCP)
-    #endif
-
-    /// Socket options that apply to all sockets.
-    public static let socket: NIOBSDSocket.OptionLevel =
-            NIOBSDSocket.OptionLevel(rawValue: SOL_SOCKET)
-}
-
 // IPv4 Options
 extension NIOBSDSocket.Option {
-    /// Add a multicast group membership.
-    public static let ip_add_membership: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IP_ADD_MEMBERSHIP)
-
-    /// Drop a multicast group membership.
-    public static let ip_drop_membership: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IP_DROP_MEMBERSHIP)
-
-    /// Set the interface for outgoing multicast packets.
-    public static let ip_multicast_if: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IP_MULTICAST_IF)
-
-    /// Control multicast loopback.
-    public static let ip_multicast_loop: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IP_MULTICAST_LOOP)
-
-    /// Control multicast time-to-live.
-    public static let ip_multicast_ttl: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IP_MULTICAST_TTL)
-
     /// Request that we are passed type of service details when receiving
     /// datagrams.
     ///
@@ -321,31 +149,6 @@ extension NIOBSDSocket.Option {
 
 // IPv6 Options
 extension NIOBSDSocket.Option {
-    /// Add an IPv6 group membership.
-    public static let ipv6_join_group: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IPV6_JOIN_GROUP)
-
-    /// Drop an IPv6 group membership.
-    public static let ipv6_leave_group: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IPV6_LEAVE_GROUP)
-
-    /// Specify the maximum number of router hops for an IPv6 packet.
-    public static let ipv6_multicast_hops: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IPV6_MULTICAST_HOPS)
-
-    /// Set the interface for outgoing multicast packets.
-    public static let ipv6_multicast_if: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IPV6_MULTICAST_IF)
-
-    /// Control multicast loopback.
-    public static let ipv6_multicast_loop: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IPV6_MULTICAST_LOOP)
-
-    /// Indicates if a socket created for the `AF_INET6` address family is
-    /// restricted to IPv6 only.
-    public static let ipv6_v6only: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: IPV6_V6ONLY)
-
     /// Request that we are passed traffic class details when receiving
     /// datagrams.
     ///
@@ -364,67 +167,6 @@ extension NIOBSDSocket.Option {
     static let ipv6_recv_pktinfo: NIOBSDSocket.Option =
         NIOBSDSocket.Option(rawValue: Posix.IPV6_RECVPKTINFO)
 }
-
-// TCP Options
-extension NIOBSDSocket.Option {
-    /// Disables the Nagle algorithm for send coalescing.
-    public static let tcp_nodelay: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: TCP_NODELAY)
-}
-
-#if os(Linux) || os(FreeBSD) || os(Android)
-extension NIOBSDSocket.Option {
-    /// Get information about the TCP connection.
-    public static let tcp_info: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: TCP_INFO)
-}
-#endif
-
-#if os(iOS) || os(macOS) || os(tvOS) || os(watchOS)
-extension NIOBSDSocket.Option {
-    /// Get information about the TCP connection.
-    public static let tcp_connection_info: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: TCP_CONNECTION_INFO)
-}
-#endif
-
-// Socket Options
-extension NIOBSDSocket.Option {
-    /// Get the error status and clear.
-    public static let so_error: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: SO_ERROR)
-
-    /// Use keep-alives.
-    public static let so_keepalive: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: SO_KEEPALIVE)
-
-    /// Linger on close if unsent data is present.
-    public static let so_linger: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: SO_LINGER)
-
-    /// Specifies the total per-socket buffer space reserved for receives.
-    public static let so_rcvbuf: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: SO_RCVBUF)
-
-    /// Specifies the receive timeout.
-    public static let so_rcvtimeo: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: SO_RCVTIMEO)
-
-    /// Allows the socket to be bound to an address that is already in use.
-    public static let so_reuseaddr: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: SO_REUSEADDR)
-}
-
-#if !os(Windows)
-extension NIOBSDSocket.Option {
-    /// Indicate when to generate timestamps.
-    public static let so_timestamp: NIOBSDSocket.Option =
-            NIOBSDSocket.Option(rawValue: SO_TIMESTAMP)
-}
-#endif
-
-/// The requested UDS path exists and has wrong type (not a socket).
-public struct UnixDomainSocketPathWrongType: Error {}
 
 /// This protocol defines the methods that are expected to be found on
 /// `NIOBSDSocket`. While defined as a protocol there is no expectation that any
@@ -533,15 +275,6 @@ protocol _BSDSocketProtocol {
                      timeout: CInt) throws -> CInt
 #endif
 
-    static func inet_ntop(af family: NIOBSDSocket.AddressFamily,
-                          src addr: UnsafeRawPointer,
-                          dst dstBuf: UnsafeMutablePointer<CChar>,
-                          size dstSize: socklen_t) throws -> UnsafePointer<CChar>?
-
-    static func inet_pton(af family: NIOBSDSocket.AddressFamily,
-                          src description: UnsafePointer<CChar>,
-                          dst address: UnsafeMutableRawPointer) throws
-
     static func sendfile(socket s: NIOBSDSocket.Handle,
                          fd: CInt,
                          offset: off_t,
@@ -585,3 +318,6 @@ protocol _BSDSocketControlMessageProtocol {
 /// If this extension is hitting a compile error, your platform is missing one
 /// of the functions defined above!
 enum NIOBSDSocketControlMessage: _BSDSocketControlMessageProtocol { }
+
+/// The requested UDS path exists and has wrong type (not a socket).
+public struct UnixDomainSocketPathWrongType: Error {}
