+++ Sources/NIO/IO.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the SwiftNIO open source project
 //
-// Copyright (c) 2017-2020 Apple Inc. and the SwiftNIO project authors
+// Copyright (c) 2017-2021 Apple Inc. and the SwiftNIO project authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
@@ -12,96 +12,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-#if os(Windows)
-import typealias WinSDK.DWORD
-#endif
-
-/// An `Error` for an IO operation.
-public struct IOError: Swift.Error {
-    @available(*, deprecated, message: "NIO no longer uses FailureDescription.")
-    public enum FailureDescription {
-        case function(StaticString)
-        case reason(String)
-    }
-
-    /// The actual reason (in an human-readable form) for this `IOError`.
-    private var failureDescription: String
-
-    @available(*, deprecated, message: "NIO no longer uses FailureDescription, use IOError.description for a human-readable error description")
-    public var reason: FailureDescription {
-        return .reason(self.failureDescription)
-    }
-
-    private enum Error {
-        #if os(Windows)
-            case windows(DWORD)
-            case winsock(CInt)
-        #endif
-        case errno(CInt)
-    }
-
-    private let error: Error
-
-    /// The `errno` that was set for the operation.
-    public var errnoCode: CInt {
-        switch self.error {
-        case .errno(let code):
-            return code
-        #if os(Windows)
-            default:
-                fatalError("IOError domain is not `errno`")
-        #endif
-        }
-    }
-
-#if os(Windows)
-    public init(windows code: DWORD, reason: String) {
-        self.error = .windows(code)
-        self.failureDescription = reason
-    }
-
-    public init(winsock code: CInt, reason: String) {
-        self.error = .winsock(code)
-        self.failureDescription = reason
-    }
-#endif
-
-    /// Creates a new `IOError``
-    ///
-    /// - parameters:
-    ///     - errorCode: the `errno` that was set for the operation.
-    ///     - reason: the actual reason (in an human-readable form).
-    public init(errnoCode code: CInt, reason: String) {
-        self.error = .errno(code)
-        self.failureDescription = reason
-    }
-
-    /// Creates a new `IOError``
-    ///
-    /// - parameters:
-    ///     - errorCode: the `errno` that was set for the operation.
-    ///     - function: The function the error happened in, the human readable description will be generated automatically when needed.
-    @available(*, deprecated, renamed: "init(errnoCode:reason:)")
-    public init(errnoCode code: CInt, function: StaticString) {
-        self.error = .errno(code)
-        self.failureDescription = "\(function)"
-    }
-}
-
-/// Returns a reason to use when constructing a `IOError`.
-///
-/// - parameters:
-///     - errorCode: the `errno` that was set for the operation.
-///     - reason: what failed
-/// - returns: the constructed reason.
-private func reasonForError(errnoCode: CInt, reason: String) -> String {
-    if let errorDescC = strerror(errnoCode) {
-        return "\(reason): \(String(cString: errorDescC)) (errno: \(errnoCode))"
-    } else {
-        return "\(reason): Broken strerror, unknown error: \(errnoCode)"
-    }
-}
-
 internal extension IOResult where T: FixedWidthInteger {
     var result: T {
         switch self {
@@ -110,16 +20,6 @@ internal extension IOResult where T: FixedWidthInteger {
         case .wouldBlock(_):
             fatalError("cannot unwrap IOResult")
         }
-    }
-}
-
-extension IOError: CustomStringConvertible {
-    public var description: String {
-        return self.localizedDescription
-    }
-
-    public var localizedDescription: String {
-        return reasonForError(errnoCode: self.errnoCode, reason: self.failureDescription)
     }
 }
